% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Batch import and align}
\usage{
batch(align = TRUE, slice = TRUE, folder = choose.dir(caption =
  "Select Save Folder"))
}
\arguments{
\item{align}{Logical. If FALSE, the raw mandible models will be imported to the global environment but not aligned. Setting this to FALSE automatically sets "slice" to FALSE.}

\item{slice}{Logical. If TRUE, the aligned point cloud will be "sliced" to isolate the mandibular angles, posterior rami, and mental eminence, which will all be plotted in a subfolder and saved as separate .xlsx files.}

\item{folder}{The save folder. If not specified, the user can interactively choose a folder.}
}
\description{
Calls align2 (and thus various other functions) in order to import and align a series of.xyz files. Files currently must be 6 columns, though the final 3 columns are not used (this will be changed at a later date).
}
\examples{
batch()
}

