% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gonialarea.R
\name{gonialarea}
\alias{gonialarea}
\title{Mandibular Angle Slice}
\usage{
gonialarea(sample, filename, folder, saveplots = TRUE)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the mandibular angles, flips to horizontal, and culls backfaces.
}

