% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xlsx2xyz.R
\name{xlsx2xyz}
\alias{xlsx2xyz}
\title{xlsx to xyz File Converter}
\usage{
xlsx2xyz(folder = choose.dir())
}
\arguments{
\item{folder}{Save folder. User can choose interactively if left blank.}
}
\description{
Batch converts .xlsx files containing 3 columns of point cloud data to .xyz.
}
\seealso{
\code{\line{xyz2xlsx}} which performs the inverse function.
}

