% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xyz2xlsx.R
\name{xyz2xlsx}
\alias{xyz2xlsx}
\title{xyz to xlsx File Converter}
\usage{
xyz2xlsx(folder = choose.dir(), saveplots = TRUE)
}
\arguments{
\item{folder}{Save folder. User can choose interactively if left blank.}
}
\description{
Batch converts .xyz files containing 3 columns of point cloud data to .xlsx.
}
\seealso{
\code{\line{xlsx2xyz}} which performs the inverse function.
}

